requirements.txt
Flask==2.3.3
pymongo==4.6.0
python-dotenv==1.0.0
flask-cors>=3.0.0
flask-jwt-extended==4.5.2
werkzeug>=2.3.0
pillow==10.1.0
pytesseract==0.3.10
pytest==7.4.2
